Separate commonjs-related types from jsg/modules.h/c++
Further cleanup in preparation for finishing the new module registry
