+++ src/workerd/jsg/BUILD.bazel
@@ -11,6 +11,7 @@ wd_cc_library(
     srcs = [
         "async-context.c++",
         "buffersource.c++",
+        "commonjs.c++",
         "compile-cache.c++",
         "dom-exception.c++",
         "inspector.c++",
@@ -31,6 +32,7 @@ wd_cc_library(
     hdrs = [
         "async-context.h",
         "buffersource.h",
+        "commonjs.h",
         "compile-cache.h",
         "dom-exception.h",
         "function.h",
